% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_scrape_helpers.R
\name{add_sleep_time}
\alias{add_sleep_time}
\title{Add sleeping time to a function.}
\usage{
add_sleep_time(.func, sleep_time)
}
\arguments{
\item{.func}{A \code{function} which is wrapped with the}

\item{sleep_time}{\code{numeric scalar} giving the sleeping time in seconds}
}
\value{
Returns a \code{function}
}
\description{

}
\examples{

}
