% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_scrape.R
\name{map2_scrape}
\alias{map2_scrape}
\alias{future_map2_scrape}
\alias{map2_scrape_dfr}
\alias{future_map2_scrape_dfr}
\title{Scrape a list of links or html files using a scrapefun using multiple inputs.}
\usage{
map2_scrape(x, y, scrapefun)

future_map2_scrape(x, y, scrapefun)

map2_scrape_dfr(x, y, scrapefun)

future_map2_scrape_dfr(x, y, scrapefun)
}
\arguments{
\item{x}{A \code{list} or \code{atomic vector}.}

\item{y}{A \code{list} or \code{atomic vector}.}

\item{scrapefun}{A \code{function} that defines the scraping logic. See
\code{vignette(scrapefuns_and_helpers)} for examples.}
}
\description{
Scrape a list of links or html files using a scrapefun using multiple inputs.
}
\section{Functions}{
\itemize{
\item \code{map2_scrape}: standard version implemented via \code{\link[purrr]{map2}}

\item \code{future_map2_scrape}: parallelised version using \code{\link[furrr]{future_map2}}

\item \code{map2_scrape_dfr}: merges results rowwise to a tibble similar to
\code{\link[purrr]{map2_dfr}}

\item \code{future_map2_scrape_dfr}: parallelised version using \code{\link[furrr]{future_map_dfr}}
}}

