% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_scrape.R
\name{map_scrape}
\alias{map_scrape}
\alias{future_map_scrape}
\alias{map_scrape_dfr}
\alias{future_map_scrape_dfr}
\title{Scrape a list of links or html files using a scrapefun}
\usage{
map_scrape(x, scrapefun)

future_map_scrape(x, scrapefun)

map_scrape_dfr(x, scrapefun)

future_map_scrape_dfr(x, scrapefun)
}
\arguments{
\item{x}{A \code{list} or \code{atomic vector}.}

\item{scrapefun}{A \code{function} that defines the scraping logic. See
\code{vignette(scrapefuns_and_helpers)} for examples.}
}
\value{
Returns a \code{list}.
}
\description{
Scrape a list of links or html files using a scrapefun
}
\section{Functions}{
\itemize{
\item \code{map_scrape}: standard version implemented via \code{\link[purrr]{map}}

\item \code{future_map_scrape}: parallelised version using \code{\link[furrr]{future_map}}

\item \code{map_scrape_dfr}: merges results rowwise to a tibble similar to
\code{\link[purrr]{map_dfr}}

\item \code{future_map_scrape_dfr}: parallelised version using \code{\link[furrr]{future_map_dfr}}
}}

