% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawler.R
\name{messagefun}
\alias{messagefun}
\title{Prints a progress message to the console as \code{YYYY-MM-DD HH:MM:SS n/N object}}
\usage{
messagefun(n, N, object)
}
\arguments{
\item{n}{\code{numeric scalar} number of current object}

\item{N}{\code{numeric scalar} total number of objects}

\item{object}{\code{character scalar or convertable to character} the object to
print in the tail of the message}
}
\value{
Prints a message to the console
}
\description{
Prints a progress message to the console as \code{YYYY-MM-DD HH:MM:SS n/N object}
}
\examples{

messagefun(1, 1000, "object")

}
