% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping_helpers.R
\name{NordVPN_quick_connect}
\alias{NordVPN_quick_connect}
\title{Invoke the "Quick Connect" command of NordVPN}
\usage{
NordVPN_quick_connect(
  nordvpn_exe = "C:\\\\Program Files\\\\NordVPN\\\\NordVPN.exe"
)
}
\arguments{
\item{nordvpn_exe}{\code{string scalar} giving the path to the NordVPN.exe.}
}
\value{
Invokes the \code{-c} command of NordVPN
}
\description{
Invoke the "Quick Connect" command of NordVPN
}
