% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping_helpers.R
\name{return_named_list}
\alias{return_named_list}
\title{Collect all local objects and return them as named list.}
\usage{
return_named_list()
}
\value{
\code{named list} of all objects in the function that don't
start with a dot
}
\description{
Makes it easier to return a named list. Recommended us is inside
an extraction function that is passed to \code{\link{do_scrape}}.
}
\examples{

# Function that uses return_named_list()
func <- function(a, b) {
  # objects prefixed with a dot get ignored
  .not_include = b
  x = a
  y = NULL
  return_named_list()
}

func(1, 2)
}
