% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_scrape_helpers.R
\name{return_result_or_message}
\alias{return_result_or_message}
\title{Wraps around a function which is returned from \code{\link[purrr]{safely}}.
Makes it return the result or the error message}
\usage{
return_result_or_message(.function)
}
\arguments{
\item{.function}{A \code{function} to be wrapped}
}
\value{
Returns a \code{function}.
}
\description{
Wraps around a function which is returned from \code{\link[purrr]{safely}}.
Makes it return the result or the error message
}
