% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapurrr-package.R
\docType{package}
\name{scrapurrr-package}
\alias{scrapurrr}
\alias{scrapurrr-package}
\title{scrapurrr: Functional Programming Style Webscraping à la purrr}
\description{
This package offers a functional programming style framework for 
    webscraping and useful tools and tutorials to overcome common webscraping 
    challenges. Check out the vignettes for webscraping workflows and infos on 
    how to write the webscraping logic for specific webscraping challenges.
}
\author{
\strong{Maintainer}: Martin G. Haas \email{haasmartin@live.de}

}
\keyword{internal}
