% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapurrr.R
\name{scrapurrr}
\alias{scrapurrr}
\title{Scrape a list of links or html files using a scrapefun}
\usage{
scrapurrr(
  ...,
  scrapefun,
  map_fun,
  delay = 0,
  attempts = 3,
  display_progress_bar = F
)
}
\arguments{
\item{...}{arguments passed to the \code{scrapefun}.}

\item{scrapefun}{A \code{function} that defines the scraping logic. See
\code{vignette(scrapefuns_and_helpers)} for examples.}

\item{map_fun}{A map_... \code{function} from \code{purrr} ord \code{furrr}.}

\item{delay}{\code{numeric scalar} giving the number of seconds to wait
before the function is called}

\item{attempts}{\code{numeric scalar} giving the number of retry attempts
if the function fails}

\item{display_progress_bar}{\code{logical scalar} indicating whether to display
a progress bar}
}
\value{
Returns a \code{list}.
}
\description{
Scrape a list of links or html files using a scrapefun
}
