% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawler.R
\name{try_and_timeout}
\alias{try_and_timeout}
\title{Tries and retries to execute the given function if it throws an error or times
out.}
\usage{
try_and_timeout(.f, ..., timeout = 15, attempts = 3, print_status_message = T)
}
\arguments{
\item{.f}{\code{function}. Function to execute.}

\item{...}{Additional arguments passed on to the mapped function.}

\item{timeout}{\code{numeric scalar}. Seconds until function is considered to
be timed out.}

\item{attempts}{\code{numeric scalar}. Number of attempts to retry execution
of \code{.f}.}

\item{print_status_message}{\code{Logical scalar}, indicating whether status
messages should be printed.}
}
\value{
Returns either a \code{character scalar} error message or the
function result.
}
\description{
Tries and retries to execute the given function if it throws an error or times
out.
}
\examples{
# Works
try_and_timeout(mean, 1:100)

# Returns try-error because of timeout
try_and_timeout(Sys.sleep, 0.2, timeout = 0.1)

# Returns error because of wrong argument
try_and_timeout(mean, "error")
}
