% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping_helpers.R
\name{view_html}
\alias{view_html}
\title{Download html and/or display it in the RStudio Viewer / Browser}
\usage{
view_html(link = NULL, html = NULL)
}
\arguments{
\item{link}{\code{string scala} (optional) Link to a webpage. It will be
downloaded via \code{\link[xml2]{read_html}}.}

\item{html}{\code{string scalar or html} containing the html to be displayed}
}
\value{
Opens the RStudio Viewer and displays the webpage or html
}
\description{
Download html and/or display it in the RStudio Viewer / Browser
}
\examples{

# Using link to webpage:
view_html("https://www.google.de")

# Using downloaded html
google = xml2::read_html("https://www.google.de")
view_html(html = google)

}
