% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscraping-package.R
\docType{package}
\name{webscraping-package}
\alias{webscraping}
\alias{webscraping-package}
\title{webscraping: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
