% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapurrr.R
\name{with_progress}
\alias{with_progress}
\title{Add progress tick \code{pb$tick()} to a function}
\usage{
with_progress(.function, pb)
}
\arguments{
\item{.function}{.function A \code{function} to which the progress should be
added}

\item{pb}{A progress bar as created by \code{\link[progress]{progress_bar}}}
}
\value{
Returns a \code{function}.
}
\description{
Add progress tick \code{pb$tick()} to a function
}
