% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawler.R
\name{try_and_timeout}
\alias{try_and_timeout}
\title{Executes a function and gives a try-error if it fails or times out}
\usage{
try_and_timeout(func, ..., timeout = 15)
}
\arguments{
\item{func}{\code{function} Function to execute}

\item{...}{\code{further arguments to function}}

\item{timeout}{\code{numeric scalar} Seconds until function is considered to
be timed out}
}
\value{
Returns either a \code{try-error} or the function result
}
\description{
Executes a function and gives a try-error if it fails or times out
}
\examples{

# Works
try_and_timeout(mean, 1:100)

# Returns try-error because of timeout
try_and_timeout(Sys.sleep, 0.2, timeout = 0.1)

# Returns error because of wrong argument
try_and_timeout(mean, "error")
}
